Add PlaylistDeleteItemViewTest to tests in playlist app#
A different class for setting up a dummy database was required for deleting, because the associated files must remain undeleted after setting up the dummy database. I named the class for this purpose SetUpDatabaseForDelete, whose only defference is it does not delete the files for the added playlist. I also extracted a method for removing associated files in playlist.views PlaylistDeleteItemView. For better readability I added a method named remove_associated_files()